LayoutTests/http/tests/security/contentSecurityPolicy/sandbox-empty.html

<script>
if (window.testRunner)
    testRunner.dumpAsText();
</script>
<meta http-equiv="Content-Security-Policy" content="sandbox">
This test passes if it doesn't alert fail.
<script>
alert('FAIL');
</script>
