LayoutTests/http/tests/security/contentSecurityPolicy/sandbox-empty-expected.txt

CONSOLE MESSAGE: Blocked script execution in 'http://127.0.0.1:8000/security/contentSecurityPolicy/sandbox-empty.html' because the document's frame is sandboxed and the 'allow-scripts' permission is not set.
This test passes if it doesn't alert fail.
